requirements.txt
py-dateutil==2.2
Scrapy==1.0.3
SQLAlchemy==1.0.9